Change redeem points to type int for testing purposes

+++ src/main/java/UStore/Redeemable.java
@@ -2,7 +2,7 @@ package UStore;
 
 public interface Redeemable {
 
-    RewardsItem redeemItem(float points);
+    RewardsItem redeemItem(int points);
 
 
 }
